requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aggdraw==1.3.12
    # via psd-tools
aiohttp==3.8.4
    # via slackclient
aiosignal==1.3.1
    # via aiohttp
amqp==2.6.1
    # via kombu
asana==0.10.3
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   psd-tools
backoff==1.6.0
    # via rudder-sdk-python
billiard==3.6.4.0
    # via celery
bleach==3.3.0
    # via -r requirements.in
cairocffi==1.2.0
    # via cairosvg
cairosvg==2.5.2
    # via -r requirements.in
celery==4.3.0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via
    #   cairocffi
    #   cryptography
chardet==3.0.4
    # via requests
charset-normalizer==3.1.0
    # via aiohttp
cryptography==3.4.7
    # via oauthlib
cssselect==1.1.0
    # via premailer
cssselect2==0.4.1
    # via cairosvg
cssutils==2.3.0
    # via premailer
defusedxml==0.7.1
    # via cairosvg
diff-match-patch==20121119
    # via -r requirements.in
django==2.2.24
    # via
    #   -r requirements.in
    #   django-jinja
    #   django-sampledatahelper
    #   django-sites
    #   django-sr
    #   easy-thumbnails
django-ipware==1.1.6
    # via -r requirements.in
django-jinja==2.7.0
    # via -r requirements.in
django-pglocks==1.0.2
    # via -r requirements.in
django-picklefield==0.3.2
    # via -r requirements.in
django-sampledatahelper==0.4.1
    # via -r requirements.in
django-sites==0.10
    # via -r requirements.in
django-sr==0.0.4
    # via -r requirements.in
djmail==1.0.1
    # via -r requirements.in
docopt==0.6.2
    # via psd-tools
easy-thumbnails==2.7.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gunicorn==19.9.0
    # via -r requirements.in
html5lib==1.1
    # via -r requirements.in
idna==2.8
    # via
    #   requests
    #   yarl
jinja2==3.0.1
    # via django-jinja
kombu==4.6.11
    # via celery
lxml==4.6.3
    # via premailer
markdown==3.3.4
    # via
    #   -r requirements.in
    #   pymdown-extensions
markupsafe==2.0.1
    # via jinja2
messagebird==2.1.0
    # via -r requirements.in
monotonic==1.6
    # via rudder-sdk-python
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
netaddr==0.7.19
    # via -r requirements.in
oauthlib[signedtoken]==3.1.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
packaging==20.9
    # via bleach
pillow==8.2.0
    # via
    #   -r requirements.in
    #   cairosvg
    #   easy-thumbnails
    #   psd-tools
premailer==3.0.1
    # via -r requirements.in
psd-tools==1.8.38
    # via -r requirements.in
psycopg2-binary==2.8.5
    # via -r requirements.in
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via -r requirements.in
pyjwt==2.1.0
    # via
    #   messagebird
    #   oauthlib
pymdown-extensions==8.1.1
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
python-dateutil==2.7.5
    # via
    #   -r requirements.in
    #   messagebird
    #   rudder-sdk-python
python-magic==0.4.15
    # via -r requirements.in
pytz==2021.1
    # via
    #   -r requirements.in
    #   celery
    #   django
    #   sampledata
raven==6.10.0
    # via -r requirements.in
redis==2.10.5
    # via -r requirements.in
requests==2.21.0
    # via
    #   -r requirements.in
    #   asana
    #   messagebird
    #   premailer
    #   requests-oauthlib
    #   rudder-sdk-python
requests-oauthlib==1.3.0
    # via
    #   -r requirements.in
    #   asana
rudder-sdk-python==1.0.0b1
    # via -r requirements.in
sampledata==0.3.7
    # via django-sampledatahelper
serpy==0.1.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asana
    #   bleach
    #   django-sampledatahelper
    #   html5lib
    #   python-dateutil
    #   rudder-sdk-python
    #   sampledata
    #   serpy
    #   webcolors
slackclient==2.9.4
    # via -r requirements.in
sqlparse==0.4.1
    # via django
tinycss2==1.1.0
    # via
    #   cairosvg
    #   cssselect2
unidecode==0.4.20
    # via -r requirements.in
urllib3==1.24.3
    # via requests
vine==1.3.0
    # via
    #   amqp
    #   celery
webcolors==1.9.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
yarl==1.9.2
    # via aiohttp
zipp==1.2.0
    # via -r requirements.in
